Use specific controller file with name matching the FE component `inc/navigation-secondary.php`
